handle shape 0 and fix comments
